sync config and refine deploy.sh

+++ deploy.sh
@@ -1,20 +1,39 @@
 #!/bin/sh
 
-#sudo echo "please use root privilege"
+#######DRY RUN FOR TEST#######
+# dry_run=0
 
+# while getopts 'n' opt; do
+#     case "$opt" in
+#         n) dry_run=1 ;;
+#         *) echo 'error in command line parsing' >&2
+#             exit 1
+#     esac
+# done
+
+# if [ "$dry_run" -eq 1 ]; then
+#     set -v
+#     set -n
+# fi
+
+#######INSTALL BEGIN#######
+
+#Install some utilities
 tools_dir="$HOME/tools"
-
-#check and set up directory "tools" 
-echo "check and set up tools directory in home"
-[ ! -d $tools_dir ] && mkdir $tools_dir
+echo "Install some utilities ..."
+[[ -d $tools_dir ]] && {
+    rm -rf $tools_dir
+}
+tar -cvf - tools | tar -xvf - -C $HOME
 
 #configure bashrc for bash
+echo "Add custom changes to .bashrc file ..."
 cp ~/.bashrc bashrc
 echo "#===========begin:user custom definition=========" >> bashrc
-echo "alias grep='grep -nr --color=auto'" >> bashrc
+echo "alias grep='grep -nr --color=auto' --exclude-dir=.ccls-cache" >> bashrc
 echo "alias rm='rm -i'" >> bashrc
 echo "source $tools_dir/aliasfile" >> bashrc
-echo "PATH=$tools_dir:$PATH" >> bashrc
+echo "PATH=$PATH:$tools_dir" >> bashrc
 
 cat >> bashrc <<'EOF'
 export LESS_TERMCAP_md=$'\E[01;31m'
@@ -26,14 +45,16 @@ export LESS_TERMCAP_us=$'\E[01;32m'
 EOF
 echo "#===========end:user custom definition=========" >> bashrc
 
-echo "******************************************************"
-diff -Nur ~/.bashrc bashrc
-echo "******************************************************"
-read -p "please confirm changes for .bashrc" tmp
-cp bashrc ~/.bashrc
+if [[ -n ${DEBUG} ]]; then
+    echo "******************************************************"
+    diff -Nur ~/.bashrc bashrc
+    echo "******************************************************"
+    read -p "please confirm changes for .bashrc" tmp
+    cp bashrc ~/.bashrc
+fi
 
-#configure .gitconfig
-echo "configure .gitconfig"
+#configure git setttings
+echo "Configure git setttings..."
 read -p "user name for git" username
 read -p "user email for git" useremail
 git config --global user.name $username
@@ -48,10 +69,16 @@ git config --global alias.rb rebase
 git config --global push.default simple
 
 #configure vim
-echo "configure vim"
-cp vimrc ~/.vimrc
-cp -rf vim ~/.vim
+echo "Configure VIM ..."
+vimrc_file="$HOME/.vimrc"
+vim_dir="$HOME/.vim"
 
-#install some utility
-echo "install some CLI utility"
-cp -rf tools/* ~/tools
+[[ -f $vimrc_file ]] && {
+    rm -rf $vimrc_file
+}
+ln -s $vim_dir/init.vim $HOME/.vimrc
+
+[[ -f $vim_dir ]] && {
+    rm -rf $vim_dir
+}
+tar -cvf - vim | tar -xvf - -C $HOME && mv vim .vim
